WIP train fine grain

+++ trainer.py
@@ -45,7 +45,7 @@ class SentimentTrainer(object):
         self.model.eval()
         loss = 0
         predictions = torch.zeros(len(dataset))
-        predictions = predictions.long()
+        predictions = predictions
         indices = torch.range(1,dataset.num_classes)
         for idx in tqdm(xrange(len(dataset)),desc='Testing epoch  '+str(self.epoch)+''):
             tree, sent, label = dataset[idx]
